% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cov_xy_fiml}
\alias{.cov_xy_fiml}
\title{Cross-covariance when y has missing values}
\usage{
.cov_xy_fiml(x_mat, y_mat, mu_x, mu_y, cov_y)
}
\arguments{
\item{x_mat}{(matrix) complete matrix}

\item{y_mat}{(matrix) matrix with missing elements}

\item{mu_x}{(vector) known mean vector of x_mat}

\item{mu_y}{(vector) known mean vector of y_mat}

\item{cov_y}{(matrix) known covariance of y_mat}
}
\value{
Complete cross-covariance
}
\description{
Cross-covariance when y has missing values
}
\keyword{internal}
