% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_e_z_zw_het}
\alias{.get_e_z_zw_het}
\title{Get conditional expectation under heteroskedastic measurement error}
\usage{
.get_e_z_zw_het(w_mat, w_se_mat, z_mat)
}
\arguments{
\item{w_mat}{(matrix) a matrix of trait estimates for n respondents
on p latent variables}

\item{w_se_mat}{(matrix) a matrix of trait estimates standard-errors
for n respondents on p latent variables}

\item{z_mat}{(matrix) a matrix of error-free covariates for n respondents
and q covariates}
}
\value{
Returns cleaned obects if all checks pass.
}
\description{
Get conditional expectation under heteroskedastic measurement error
}
\keyword{internal}
