% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.negll_fiml}
\alias{.negll_fiml}
\title{Negative log-likelihood for MVN}
\usage{
.negll_fiml(theta, x_mat)
}
\arguments{
\item{theta}{(vector) parameter candidates}

\item{x_mat}{(matrix) matrix with missing elements}
}
\value{
Negative log-likelihood for MVN
}
\description{
Negative log-likelihood for MVN
}
\keyword{internal}
