% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.optim_cov_fiml}
\alias{.optim_cov_fiml}
\title{Optimisation function for MVN FIML}
\usage{
.optim_cov_fiml(x_mat)
}
\arguments{
\item{x_mat}{(matrix) matrix with missing elements}
}
\value{
Result from optim
}
\description{
Optimisation function for MVN FIML
}
\keyword{internal}
