% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.prep_score}
\alias{.prep_score}
\title{Prepare sum-score, EAP or regression factor scores for calibration}
\usage{
.prep_score(
  w_mat,
  rel_vec = NULL,
  w_se_vec = NULL,
  w_se_mat = NULL,
  standard = FALSE
)
}
\arguments{
\item{w_mat}{(matrix) a matrix of trait estimates for n respondents
on p latent variables}

\item{rel_vec}{(vector) a vector of marginal reliability
for p latent variables}

\item{w_se_vec}{(vector) a vector of standard errors of measurement
for p latent variables}

\item{w_se_mat}{(matrix) a matrix of trait estimates standard-errors
for n respondents on p latent variables}

\item{standard}{(logical) Only relevant when passing reliability.
If TRUE, attempt to return the standardised version of the calibrated scores
If FALSE (default), do not attempt standardisation.}
}
\value{
Returns rescaled scores.
}
\description{
Prepare sum-score, EAP or regression factor scores for calibration
}
\keyword{internal}
