% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.score_mvn}
\alias{.score_mvn}
\title{Score function for FIML of MVN}
\usage{
.score_mvn(theta, x_mat)
}
\arguments{
\item{theta}{(vector) parameter candidates}

\item{x_mat}{(matrix) matrix with missing elements}
}
\value{
Gradients of mean and covariance
}
\description{
Score function for FIML of MVN
}
\keyword{internal}
