% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.validate_calib_me}
\alias{.validate_calib_me}
\title{Validate inputs for rccme_calib_me}
\usage{
.validate_calib_me(
  w_mat,
  rel_vec = NULL,
  w_se_vec = NULL,
  w_se_mat = NULL,
  z_mat
)
}
\arguments{
\item{w_mat}{(matrix) a matrix of trait estimates for n respondents
on p latent variables}

\item{rel_vec}{(vector) a vector of marginal reliability
for p latent variables}

\item{w_se_vec}{(vector) a vector of standard errors of measurement
for p latent variables}

\item{w_se_mat}{(matrix) a matrix of trait estimates standard-errors
for n respondents on p latent variables}

\item{z_mat}{(matrix) a matrix of error-free covariates for n respondents
and q covariates}
}
\value{
Returns cleaned obects if all checks pass.
}
\description{
Validate inputs for rccme_calib_me
}
\keyword{internal}
