% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estress}
\alias{estress}
\title{ESTRESS data}
\format{
\subsection{\code{estress}}{

A data frame with 7 variables, 262 cases:
\describe{
\item{tenure}{Company Tenure}
\item{estress}{Economic stress}
\item{affect}{Depressed affect}
\item{withdraw}{Withdrawal intentions}
\item{sex}{Male (1) or Female (0)}
\item{age}{age}
\item{ese}{Entrepreneurial self efficacy}
}
}
}
\source{
Pollack, J., VanEpps, E. M., & Hayes, A. F. (2012).
The moderating role of social ties on entrepreneurs' depressed
affect and withdrawal intentions in response to economic stress.
Journal of Organizational Behavior, 33, 789-810.

Hayes, A. (2017). Introduction to mediation, moderation,
and conditional process analysis: A regression-based approach
(2nd edn, p. 692). Guilford Press.

Moon K (2023). \emph{processR: Implementation of the 'PROCESS' Macro}.
R package version 0.2.8, commit c6c817b5b716f2615b9814cd5022aef1cbc5c09d,
\url{https://github.com/cardiomoon/processR}.
}
\usage{
estress
}
\description{
ESTRESS data
}
\keyword{datasets}
