% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccme_calib_me.R
\name{rccme_calib_me}
\alias{rccme_calib_me}
\title{Calibrate scores under measurement error}
\usage{
rccme_calib_me(
  w_mat,
  rel_vec = NULL,
  w_se_vec = NULL,
  w_se_mat = NULL,
  z_mat = matrix(0, nrow = nrow(w_mat), ncol = 0),
  rescale = TRUE,
  standard = FALSE
)
}
\arguments{
\item{w_mat}{(matrix) a matrix of trait estimates for n respondents
on p latent variables}

\item{rel_vec}{(vector) a vector of marginal reliability
for p latent variables}

\item{w_se_vec}{(vector) a vector of standard errors of measurement
for p latent variables}

\item{w_se_mat}{(matrix) a matrix of trait estimates standard-errors
for n respondents on p latent variables}

\item{z_mat}{(matrix) a matrix of error-free covariates for n respondents
and q covariates}

\item{rescale}{(logical) Should the trait estimates and their
standard errors be re-scaled? Default is TRUE. The variables are
rescaled under the assumption that the trait estimates were not
conditioned on any background variables and that the marginal
variance of the latent trait is 1. This is true for standardised
latent variables in CFA or the default prior variance assumption
of 1 in IRT score estimates. This rescaling ensures the trait
coefficients are correct for the standardised traits.}

\item{standard}{(logical) Only relevant when passing reliability.
If TRUE, attempt to return the standardised version of the calibrated scores
If FALSE (default), do not attempt standardisation.}
}
\value{
Calibrated trait estimates.
}
\description{
Calibrate scores under measurement error
}
